Yoda kill! Yoda control! Bug Squashing too!
- Yoda control will look in the current directory for a configuration
file, if none is found it’ll keep going up to it’s parent folder to
find a configuration file until it gets to ‘/‘ and then it will bail.
- By default, yoda control will docker exec on the last container in
the configuration file(which I consider to be the meat and potatoes)
and perform a bash command.
- Special keyword control added to .yoda configuration file. If you
provide an array/string it will perform said commands inside the
configuration
- I believe I fixed the do/do not/no worries messaging based on docker
rm and docker kill. It appears to be exiting properly, and it also
appears to be warning properly now too.
- Added yoda kill command, which will kill all of the running containers

Change-Id: I8e405b856e6de15ac567ed304d4fceb8cd6a00bb